whatsapp_report_agent/requirements.txt
fastapi
uvicorn
openpyxl
python-dotenv
requests
pydantic
apscheduler
openai
